requirements.txt
streamlit>=1.30.0
pandas>=2.2.0
plotly>=5.20.0
python-dotenv>=1.0.0
databricks-sql-connector>=0.19.0
